refactor: 음성 입출력 관련 state 파일 src/stores로 이동
음성 입출력 관련 state 파일을 src/stores로 이동했습니다.
