+++ frontend/src/stores/stateSelectedSpeaker.tsx
